+++ R/deps.R
@@ -1,15 +1,17 @@
 #' Get the semantic.dashboard dependencies
 #'
-#' @description To add dependencies in the future follow the \code{\link[htmltools]{htmlDependency}} help.
+#' @description To add dependencies in the future follow the \code{\link[htmltools]{htmlDependency}}
+#'   help.
 #'
 #' @return semantic.dashboard dependencies
+#' @keywords internal
 get_dashboard_dependencies <- function() {
   list(
     htmlDependency(
       "semantic.dashboard",
       as.character(utils::packageVersion("semantic.dashboard")),
       c(file = system.file(package = "semantic.dashboard")),
-      script = "semantic.dashboard.min.js",
+      script = c("semantic.dashboard.min.js", "updateTabItems.js"),
       stylesheet = "semantic.dashboard.min.css"
     )
   )