+++ inst/updateTabItems.js
@@ -1,0 +1,6 @@
+/* update_tab_items */
+Shiny.addCustomMessageHandler('update_tab',
+  function(custom_tab) {
+    $(`#uisidebar [data-value='${custom_tab}']`).click();
+  }
+)